requirements.txt
fastapi
uvicorn
openai
python-multipart
PyMuPDF
python-dotenv

